mios_image: Add support for 64bit ELF

+++ host/mios_image.h
@@ -8,9 +8,9 @@ typedef struct mios_image {
   uint8_t app_version[21];  // 20 bytes SHA1 + 1 byte dirty flag
   uint8_t mios_version[21]; // 20 bytes SHA1 + 1 byte dirty flag
   uint8_t buildid[20];      // 20 byte SHA1 from toolchain
-  uint32_t buildid_paddr;   // Physical address of buildid on chip
+  uint64_t buildid_paddr;   // Physical address of buildid on chip
 
-  uint32_t load_addr;       // Where to start writing image to flash
+  uint64_t load_addr;       // Where to start writing image to flash
   size_t image_size;        // Number of bytes in image[]
   uint8_t image[0];         // Actual bytes
 } mios_image_t;
